pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lab-orchestrator"
version = "0.1.0"
description = "Lab Platform Orchestrator - Central coordination service"
authors = [{name = "Farah Orfaly"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "Jinja2==3.1.4",
    "paho-mqtt==1.6.1",
    "apscheduler==3.10.4",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
]

[project.scripts]
lab-orchestrator = "lab_orchestrator.host:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
